refactor(repository): deduplicate copy/pasted code
