+++ specifyweb/context/collection_resources.py
@@ -2,76 +2,13 @@ from specifyweb.context.app_resource import get_usertype
 from specifyweb.specify.models import Spappresource, Spappresourcedir
 from specifyweb.specify.views import openapi
 from specifyweb.context.resources import Resource, Resources
+from specifyweb.context.user_resources import get_resources_endpoint_schema, get_resource_endpoint_schema
 
-collection_resources = openapi(schema={
-    "get": {
-        "responses": {
-            "200": {
-                "description": "Returns list of public app resources in the logged in collection.",
-                "content": {
-                    "application/json": {
-                        "schema": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "properties": {
-                                    "id": { "type": "integer", "description": "The appresource id." },
-                                    "name": { "type": "string", "description": "The appresource name." },
-                                    "mimetype": { "type": "string" },
-                                    "metadata": { "type": "string" },
-                                },
-                                'required': ['id', 'name', 'mimetype', 'metadata'],
-                                'additionalProperties': False
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    },
-    "post": {
-        "requestBody": {
-            "required": True,
-            "description": "Creates appresource in the logged in collection owned by the logged in user.",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "type": "object",
-                        "properties": {
-                            "name": { "type": "string", "description": "The appresource name." },
-                            "mimetype": { "type": "string" },
-                            "metadata": { "type": "string" },
-                            "data": { "type": "string", "description": "The data to be stored in the appresource." },
-                        },
-                        'required': ['name', 'mimetype', 'metadata', 'data'],
-                        'additionalProperties': False
-                    }
-                }
-            }
-        },
-        "responses": {
-            "201": {
-                "description": "The user resource was created.",
-                "content": {
-                    "application/json": {
-                        "schema": {
-                            "type": "object",
-                            "properties": {
-                                "id": { "type": "integer", "description": "The appresource id." },
-                                "name": { "type": "string", "description": "The appresource name." },
-                                "mimetype": { "type": "string" },
-                                "metadata": { "type": "string" },
-                                "data": { "type": "string", "description": "The data to be stored in the appresource." },
-                            },
-                            'required': ['name', 'mimetype', 'metadata', 'data'],
-                            'additionalProperties': False
-                        }
-                    }
-                }
-            }
-        }
-    }
-})(Resources.as_view(_spappresourcedirfilter= lambda request: {
+collection_resources = openapi(schema=get_resources_endpoint_schema(
+    description_get="Returns list of public app resources in the logged in collection.",
+    description_create="Creates appresource in the logged in collection.",
+    description_created="The collection resource was created.",
+))(Resources.as_view(_spappresourcedirfilter= lambda request: {
                 'ispersonal': False,
                 'specifyuser__isnull': True,
                 'usertype__isnull': True,
@@ -86,60 +23,10 @@ collection_resources = openapi(schema={
 }))
 
 
-collection_resource = openapi(schema={
-    "get": {
-        "responses": {
-            "200": {
-                "description": "The public app resource of the given id in the logged in collection ",
-                "content": {
-                    "application/json": {
-                        "schema": {
-                            "type": "object",
-                            "properties": {
-                                "id": { "type": "integer", "description": "The appresource id." },
-                                "name": { "type": "string", "description": "The appresource name." },
-                                "mimetype": { "type": "string" },
-                                "metadata": { "type": "string" },
-                                "data": { "type": "string", "description": "The data to be stored in the appresource." },
-                            },
-                            'required': ['id', 'name', 'mimetype', 'metadata', 'data'],
-                            'additionalProperties': False
-                        }
-                    }
-                }
-            }
-        }
-    },
-    "put": {
-        "requestBody": {
-            "required": True,
-            "description": "Updates the appresource with the given id in the logged in collection",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "type": "object",
-                        "properties": {
-                            "name": { "type": "string", "description": "The appresource name." },
-                            "mimetype": { "type": "string" },
-                            "metadata": { "type": "string" },
-                            "data": { "type": "string", "description": "The data to be stored in the appresource." },
-                        },
-                        'required': ['name', 'mimetype', 'metadata', 'data'],
-                        'additionalProperties': False
-                    }
-                }
-            }
-        },
-        "responses": {
-            "204": { "description": "The resource was updated.", },
-        }
-    },
-    "delete": {
-        "responses": {
-            "204": {"description": "The resource was deleted.",}
-        }
-    }
-})(Resource.as_view(_spappresourcefilter= lambda request: {
+collection_resource = openapi(schema=get_resource_endpoint_schema(
+    description_get="The public app resource of the given id in the logged in collection",
+    description_update="Updates the appresource with the given id in the logged in collection"
+))(Resource.as_view(_spappresourcefilter= lambda request: {
             'spappresourcedir__ispersonal': False,
 }))
