+++ specifyweb/frontend/js_src/lib/components/DataModel/legacyTypes.ts
@@ -14,7 +14,7 @@ import type {
 } from './helperTypes';
 import type { Collection, SpecifyTable } from './specifyTable';
 
-type AllowedFields<SCHEMA extends AnySchema> =
+type SchemaFields<SCHEMA extends AnySchema> =
   | keyof CommonFields
   | keyof SCHEMA['fields']
   | keyof SCHEMA['toManyDependent']
@@ -22,13 +22,17 @@ type AllowedFields<SCHEMA extends AnySchema> =
   | keyof SCHEMA['toOneDependent']
   | keyof SCHEMA['toOneIndependent'];
 
-type AllowedValues<SCHEMA extends AnySchema> = (CommonFields &
+type SchemaFieldValue<
+  SCHEMA extends AnySchema,
+  FIELD_NAME extends SchemaFields<SCHEMA>
+> = (CommonFields &
   IR<never> &
   SCHEMA['fields'] &
   SCHEMA['toManyDependent'] &
   SCHEMA['toManyIndependent'] &
   SCHEMA['toOneDependent'] &
-  SCHEMA['toOneIndependent'])[AllowedFields<SCHEMA>];
+  SCHEMA['toOneIndependent'])[FIELD_NAME];
+
 /*
  * FEATURE: need to improve the typing to handle the following:
  *    Dynamic references
@@ -58,20 +62,20 @@ export type SpecifyResource<SCHEMA extends AnySchema> = {
    */
   /* eslint-disable @typescript-eslint/method-signature-style */
   get<
-    FIELD_NAME extends AllowedFields<SCHEMA>,
-    VALUE extends AllowedValues<SCHEMA>,
+    FIELD_NAME extends SchemaFields<SCHEMA>,
+    VALUE extends SchemaFieldValue<SCHEMA, FIELD_NAME>
   >(
     fieldName: FIELD_NAME
     // eslint-disable-next-line functional/prefer-readonly-type
   ): [VALUE] extends [never]
     ? never
     : VALUE extends AnySchema
-      ? VALUE extends null
-        ? string | null
-        : string
-      : VALUE extends RA<AnySchema>
-        ? string
-        : VALUE;
+    ? VALUE extends null
+      ? string | null
+      : string
+    : VALUE extends RA<AnySchema>
+    ? string
+    : VALUE;
   // Case-insensitive fetch of a -to-one resource
   rgetPromise<
     FIELD_NAME extends
@@ -79,7 +83,7 @@ export type SpecifyResource<SCHEMA extends AnySchema> = {
       | keyof SCHEMA['toOneIndependent'],
     VALUE = (IR<never> &
       SCHEMA['toOneDependent'] &
-      SCHEMA['toOneIndependent'])[FIELD_NAME],
+      SCHEMA['toOneIndependent'])[FIELD_NAME]
   >(
     fieldName: FIELD_NAME,
     prePopulate?: boolean
@@ -96,7 +100,7 @@ export type SpecifyResource<SCHEMA extends AnySchema> = {
       | keyof SCHEMA['toOneIndependent'],
     VALUE = (IR<never> &
       SCHEMA['toOneDependent'] &
-      SCHEMA['toOneIndependent'])[FIELD_NAME],
+      SCHEMA['toOneIndependent'])[FIELD_NAME]
   >(
     fieldName: FIELD_NAME,
     options?: {
@@ -115,14 +119,14 @@ export type SpecifyResource<SCHEMA extends AnySchema> = {
     FIELD_NAME extends keyof (SCHEMA['toManyDependent'] &
       SCHEMA['toManyIndependent']),
     VALUE extends (SCHEMA['toManyDependent'] &
-      SCHEMA['toManyIndependent'])[FIELD_NAME],
+      SCHEMA['toManyIndependent'])[FIELD_NAME]
   >(
     fieldName: FIELD_NAME,
     filters?: CollectionFetchFilters<VALUE[number]>
   ): Promise<Collection<VALUE[number]>>;
   set<
-    FIELD_NAME extends AllowedFields<SCHEMA>,
-    VALUE extends AllowedValues<SCHEMA>,
+    FIELD_NAME extends SchemaFields<SCHEMA>,
+    VALUE extends SchemaFieldValue<SCHEMA, FIELD_NAME>
   >(
     fieldName: FIELD_NAME,
     value: readonly [VALUE] extends readonly [never]
@@ -140,11 +144,11 @@ export type SpecifyResource<SCHEMA extends AnySchema> = {
                   | RA<SerializedResource<VALUE[number]>>
                   | RA<SpecifyResource<VALUE[number]>>
               : null extends VALUE
-                ?
-                    | SerializedResource<Exclude<VALUE, null>>
-                    | SpecifyResource<Exclude<VALUE, null>>
-                    | null
-                : SerializedResource<VALUE> | SpecifyResource<VALUE>),
+              ?
+                  | SerializedResource<Exclude<VALUE, null>>
+                  | SpecifyResource<Exclude<VALUE, null>>
+                  | null
+              : SerializedResource<VALUE> | SpecifyResource<VALUE>),
     options?: { readonly silent: boolean }
   ): SpecifyResource<SCHEMA>;
   // Not type safe
