+++ specifyweb/frontend/js_src/lib/components/Molecules/Tabs.tsx
@@ -2,27 +2,33 @@ import { Tab } from '@headlessui/react';
 import React from 'react';
 
 import type { GetSet, IR } from '../../utils/types';
-import { className } from '../Atoms/className';
+import { radioButtonClassName } from '../AppResources/Filters';
 import { ErrorBoundary } from '../Errors/ErrorBoundary';
 
 export function Tabs({
   tabs,
   index: [currentIndex, handleChange],
+  variant = 'separated',
 }: {
   readonly tabs: IR<JSX.Element>;
   readonly index: GetSet<number>;
+  readonly variant?: 'combined' | 'separated';
 }): JSX.Element {
   return (
     <Tab.Group selectedIndex={currentIndex} onChange={handleChange}>
       <Tab.List
-        // Don't display tabs if there is only one tab
-        className={`flex flex-wrap gap-2 ${
-          Object.keys(tabs).length === 1 ? 'sr-only' : ''
-        }`}
+        className={`
+          inline-flex w-fit flex-wrap gap-2 rounded
+          ${variant === 'separated' ? '' : 'bg-[color:var(--form-background)]'}
+          ${
+            // Don't display tabs if there is only one tab
+            Object.keys(tabs).length === 1 ? 'sr-only' : ''
+          }
+        `}
       >
         {Object.keys(tabs).map((label, index) => (
           <Tab
-            className={`${className.niceButton} ${className.infoButton}`}
+            className={radioButtonClassName(currentIndex === index)}
             key={index}
             /**
              * HeadlessUI does not trigger onChange on click on current tab.
@@ -30,7 +36,9 @@ export function Tabs({
              * if the option IS current.
              */
             onClick={
-              currentIndex === index ? () => handleChange(index) : undefined
+              currentIndex === index
+                ? (): void => handleChange(index)
+                : undefined
             }
           >
             {label}
