+++ specifyweb/specify/build_models.py
@@ -2,29 +2,11 @@ from django.db import models
 from django.db.models.signals import pre_delete
 
 from specifyweb.businessrules.exceptions import AbortSave
+from specifyweb.specify.config import orderings
 from . import model_extras
 from .model_timestamp import save_auto_timestamp_field_with_override
 
 appname = __name__.split('.')[-2]
-
-# REFACTOR: generate this on the fly based on presence of
-#    ordinal/ordernumber/position etc field in a table. This way it will
-#    automatically work for any newly added table
-orderings = {
-    'Picklistitem': ('ordinal', ),
-    'Recordsetitem': ('recordid', ),
-    'Spqueryfield': ('position', ),
-    'Determination': ('-iscurrent',),
-    'Author': ('ordernumber',),
-    'Collector': ('ordernumber',),
-    'AgentSpecialty': ('ordernumber',),
-    'Determiner': ('ordernumber',),
-    'Extractor': ('ordernumber',),
-    'FieldNotebookPageSet': ('ordernumber',),
-    'FundingAgent': ('ordernumber',),
-    'GroupPerson': ('ordernumber',),
-    'PcrPerson': ('ordernumber',),
-}
 
 def make_model(module, table, datamodel):
     """Returns a Django model class based on the
