+++ specifyweb/specify/config.py
@@ -1,0 +1,154 @@
+sp6_dependent_fields = {
+    'Accession.accessionagents',
+    'Accession.accessionauthorizations',
+    'Accession.addressofrecord',
+    'Agent.addresses',
+    'Agent.agentgeographies',
+    'Agent.agentspecialties',
+    'Agent.groups',
+    'Agent.identifiers',
+    'Agent.variants',
+    'Borrow.addressofrecord',
+    'Borrow.borrowagents',
+    'Borrow.borrowmaterials',
+    'Borrow.shipments',
+    'Borrowmaterial.borrowreturnmaterials',
+    'Collectingevent.collectingeventattribute',
+    'Collectingevent.collectingeventattrs',
+    'Collectingevent.collectingeventauthorizations',
+    'Collectingevent.collectors',
+    'Collectingtrip.collectingtripattribute',
+    'Collectingtrip.collectingtripauthorizations',
+    'Collectingtrip.fundingagents',
+    'Collectionobject.collectionobjectattribute',
+    'Collectionobject.collectionobjectattrs',
+    'Collectionobject.collectionobjectcitations',
+    'Collectionobject.conservdescriptions',
+    'Collectionobject.determinations',
+    'Collectionobject.dnasequences',
+    'Collectionobject.exsiccataitems',
+    'CollectionObject.leftsiderels',
+    'Collectionobject.otheridentifiers',
+    'Collectionobject.preparations',
+    'Collectionobject.collectionobjectproperties',
+    'CollectionObject.rightsiderels',
+    'Collectionobject.treatmentevents',
+    'Collectionobject.voucherrelationships',
+    'Commonnametx.citations',
+    'Conservdescription.events',
+    'Deaccession.deaccessionagents',
+    'Determination.determinationcitations',
+    'Determination.determiners',
+    'Disposal.disposalagents',
+    'Disposal.disposalpreparations',
+    'Dnasequence.dnasequencingruns',
+    'Dnasequencingrun.citations',
+    'Exchangein.exchangeinpreps',
+    'Exchangein.addressofrecord',
+    'Exchangeout.exchangeoutpreps',
+    'Exchangeout.addressofrecord',
+    'Exsiccata.exsiccataitems',
+    'Fieldnotebook.pagesets',
+    'Fieldnotebookpageset.pages',
+    'Geographytreedef.treedefitems',
+    'Geologictimeperiodtreedef.treedefitems',
+    'Gift.addressofrecord',
+    'Gift.giftagents',
+    'Gift.giftpreparations',
+    'Gift.shipments',
+    'Latlonpolygon.points',
+    'lithostrattreedef.treedefitems',
+    'Loan.addressofrecord',
+    'Loan.loanagents',
+    'Loan.loanpreparations',
+    'Loan.shipments',
+    'Loanpreparation.loanreturnpreparations',
+    'Locality.geocoorddetails',
+    'Locality.latlonpolygons',
+    'Locality.localitycitations',
+    'Locality.localitydetails',
+    'Locality.localitynamealiass',
+    'Materialsample.dnasequences',
+    'Picklist.picklistitems',
+    'Preparation.materialsamples',
+    'Preparation.preparationattribute',
+    'Preparation.preparationattrs',
+    'Preparation.preparationproperties',
+    'Preptype.attributedefs',
+    'Referencework.authors',
+    'Repositoryagreement.addressofrecord',
+    'Repositoryagreement.repositoryagreementagents',
+    'Repositoryagreement.repositoryagreementauthorizations',
+    'Spquery.fields',
+    'Storagetreedef.treedefitems',
+    'Taxon.commonnames',
+    'Taxon.taxoncitations',
+    'Taxon.taxonattribute',
+    'Taxontreedef.treedefitems',
+    'Workbench.workbenchtemplate',
+    'Workbenchtemplate.workbenchtemplatemappingitems',
+}
+
+
+sp6_system_tables = {
+    'Attachment',
+    'Attachmentimageattribute',
+    'Attachmentmetadata',
+    'Attachmenttag',
+    'Attributedef',
+    'Autonumberingscheme',
+    'Datatype',
+    'Morphbankview',
+    'Picklist',
+    'Picklistitem',
+    'Recordset',
+    'Recordsetitem',
+    'Spappresource',
+    'Spappresourcedata',
+    'Spappresourcedir',
+    'Spauditlogfield',
+    'Spexportschema',
+    'Spexportschemaitem',
+    'Spexportschemaitemmapping',
+    'Spexportschemamapping',
+    'Spfieldvaluedefault',
+    'Splocalecontainer',
+    'Splocalecontaineritem',
+    'Splocaleitemstr',
+    'Sppermission',
+    'Spprincipal',
+    'Spquery',
+    'Spqueryfield',
+    'Spreport',
+    'Sptasksemaphore',
+    'Spversion',
+    'Spviewsetobj',
+    'Spvisualquery',
+    'Specifyuser',
+    'Workbench',
+    'Workbenchdataitem',
+    'Workbenchrow',
+    'Workbenchrowexportedrelationship',
+    'Workbenchrowimage',
+    'Workbenchtemplate',
+    'Workbenchtemplatemappingitem',
+}
+
+# REFACTOR: generate this on the fly based on presence of
+#    ordinal/ordernumber/position etc field in a table. This way it will
+#    automatically work for any newly added table
+orderings = {
+    'Picklistitem': ('ordinal', ),
+    'Recordsetitem': ('recordid', ),
+    'Spqueryfield': ('position', ),
+    'Determination': ('-iscurrent',),
+    'Author': ('ordernumber',),
+    'Collector': ('ordernumber',),
+    'AgentSpecialty': ('ordernumber',),
+    'Determiner': ('ordernumber',),
+    'Extractor': ('ordernumber',),
+    'FieldNotebookPageSet': ('ordernumber',),
+    'FundingAgent': ('ordernumber',),
+    'GroupPerson': ('ordernumber',),
+    'PcrPerson': ('ordernumber',),
+}
