+++ specifyweb/specify/load_datamodel.py
@@ -8,6 +8,8 @@ logger = logging.getLogger(__name__)
 from django.conf import settings # type: ignore
 from django.utils.translation import gettext as _
 
+from specifyweb.specify.config import sp6_dependent_fields, sp6_system_tables
+
 class DoesNotExistError(Exception):
     pass
 
@@ -387,139 +389,3 @@ def flag_system_tables(datamodel: Datamodel) -> None:
             table.system = True
         if table.name.endswith('treedef') or table.name.endswith('treedefitem'):
             table.system = True
-
-sp6_dependent_fields = {
-    'Accession.accessionagents',
-    'Accession.accessionauthorizations',
-    'Accession.addressofrecord',
-    'Agent.addresses',
-    'Agent.agentgeographies',
-    'Agent.agentspecialties',
-    'Agent.groups',
-    'Agent.identifiers',
-    'Agent.variants',
-    'Borrow.addressofrecord',
-    'Borrow.borrowagents',
-    'Borrow.borrowmaterials',
-    'Borrow.shipments',
-    'Borrowmaterial.borrowreturnmaterials',
-    'Collectingevent.collectingeventattribute',
-    'Collectingevent.collectingeventattrs',
-    'Collectingevent.collectingeventauthorizations',
-    'Collectingevent.collectors',
-    'Collectingtrip.collectingtripattribute',
-    'Collectingtrip.collectingtripauthorizations',
-    'Collectingtrip.fundingagents',
-    'Collectionobject.collectionobjectattribute',
-    'Collectionobject.collectionobjectattrs',
-    'Collectionobject.collectionobjectcitations',
-    'Collectionobject.conservdescriptions',
-    'Collectionobject.determinations',
-    'Collectionobject.dnasequences',
-    'Collectionobject.exsiccataitems',
-    'CollectionObject.leftsiderels',
-    'Collectionobject.otheridentifiers',
-    'Collectionobject.preparations',
-    'Collectionobject.collectionobjectproperties',
-    'CollectionObject.rightsiderels',
-    'Collectionobject.treatmentevents',
-    'Collectionobject.voucherrelationships',
-    'Commonnametx.citations',
-    'Conservdescription.events',
-    'Deaccession.deaccessionagents',
-    'Determination.determinationcitations',
-    'Determination.determiners',
-    'Disposal.disposalagents',
-    'Disposal.disposalpreparations',
-    'Dnasequence.dnasequencingruns',
-    'Dnasequencingrun.citations',
-    'Exchangein.exchangeinpreps',
-    'Exchangein.addressofrecord',
-    'Exchangeout.exchangeoutpreps',
-    'Exchangeout.addressofrecord',
-    'Exsiccata.exsiccataitems',
-    'Fieldnotebook.pagesets',
-    'Fieldnotebookpageset.pages',
-    'Geographytreedef.treedefitems',
-    'Geologictimeperiodtreedef.treedefitems',
-    'Gift.addressofrecord',
-    'Gift.giftagents',
-    'Gift.giftpreparations',
-    'Gift.shipments',
-    'Latlonpolygon.points',
-    'lithostrattreedef.treedefitems',
-    'Loan.addressofrecord',
-    'Loan.loanagents',
-    'Loan.loanpreparations',
-    'Loan.shipments',
-    'Loanpreparation.loanreturnpreparations',
-    'Locality.geocoorddetails',
-    'Locality.latlonpolygons',
-    'Locality.localitycitations',
-    'Locality.localitydetails',
-    'Locality.localitynamealiass',
-    'Materialsample.dnasequences',
-    'Picklist.picklistitems',
-    'Preparation.materialsamples',
-    'Preparation.preparationattribute',
-    'Preparation.preparationattrs',
-    'Preparation.preparationproperties',
-    'Preptype.attributedefs',
-    'Referencework.authors',
-    'Repositoryagreement.addressofrecord',
-    'Repositoryagreement.repositoryagreementagents',
-    'Repositoryagreement.repositoryagreementauthorizations',
-    'Spquery.fields',
-    'Storagetreedef.treedefitems',
-    'Taxon.commonnames',
-    'Taxon.taxoncitations',
-    'Taxon.taxonattribute',
-    'Taxontreedef.treedefitems',
-    'Workbench.workbenchtemplate',
-    'Workbenchtemplate.workbenchtemplatemappingitems',
-}
-
-
-sp6_system_tables = {
-    'Attachment',
-    'Attachmentimageattribute',
-    'Attachmentmetadata',
-    'Attachmenttag',
-    'Attributedef',
-    'Autonumberingscheme',
-    'Datatype',
-    'Morphbankview',
-    'Picklist',
-    'Picklistitem',
-    'Recordset',
-    'Recordsetitem',
-    'Spappresource',
-    'Spappresourcedata',
-    'Spappresourcedir',
-    'Spauditlogfield',
-    'Spexportschema',
-    'Spexportschemaitem',
-    'Spexportschemaitemmapping',
-    'Spexportschemamapping',
-    'Spfieldvaluedefault',
-    'Splocalecontainer',
-    'Splocalecontaineritem',
-    'Splocaleitemstr',
-    'Sppermission',
-    'Spprincipal',
-    'Spquery',
-    'Spqueryfield',
-    'Spreport',
-    'Sptasksemaphore',
-    'Spversion',
-    'Spviewsetobj',
-    'Spvisualquery',
-    'Specifyuser',
-    'Workbench',
-    'Workbenchdataitem',
-    'Workbenchrow',
-    'Workbenchrowexportedrelationship',
-    'Workbenchrowimage',
-    'Workbenchtemplate',
-    'Workbenchtemplatemappingitem',
-}
