+++ specifyweb/specify/sp7_build_datamodel.py
@@ -4,6 +4,8 @@ import logging
 from django.conf import settings # type: ignore
 from django.utils.translation import gettext as _
 
+from specifyweb.specify.config import sp6_dependent_fields, sp6_system_tables
+
 logger = logging.getLogger(__name__)
 
 TAB1 = '    '
